@@ -1,9 +1,13 @@
 module github.com/jmoiron/sqlx
 
-go 1.10
+go 1.23.4
 
 require (
-	github.com/go-sql-driver/mysql v1.6.0
-	github.com/lib/pq v1.2.0
-	github.com/mattn/go-sqlite3 v1.14.6
+	github.com/go-sql-driver/mysql v1.8.1
+	github.com/goccy/go-json v0.10.4
+	github.com/goccy/go-reflect v1.2.0
+	github.com/lib/pq v1.10.9
+	github.com/mattn/go-sqlite3 v1.14.24
 )
+
+require filippo.io/edwards25519 v1.1.0 // indirect
