@@ -4,10 +4,10 @@ import (
 	"bytes"
 	"compress/gzip"
 	"database/sql/driver"
-	"encoding/json"
 	"errors"
+	"io"
 
-	"io/ioutil"
+	"github.com/goccy/go-json"
 )
 
 // GzippedText is a []byte which transparently gzips data being submitted to
@@ -43,7 +43,7 @@ func (g *GzippedText) Scan(src interface{}) error {
 		return err
 	}
 	defer reader.Close()
-	b, err := ioutil.ReadAll(reader)
+	b, err := io.ReadAll(reader)
 	if err != nil {
 		return err
 	}
